Add ESP32-S3 platform scaffold with stub drivers, build system integration, and split platform-specific tools into platform mk directories (#15016)
* Add ESP32-S3 platform scaffold with stub drivers, build system integration, and split platform-specific tools into platform mk directories

* Fix ESP32 submodule registration, redirect ESP tools to tools/espressif, and cache platform SDKs in CI

* Split platform SDK caches in CI to restore only the submodule needed per matrix target, fix ESP32 branch ref and toolchain path resolution

* Fix ESP32 build: resolve duplicate symbols, add missing stubs, fix linker script, correct toolchain path, and fix sign-compare/pointer-type warnings for xtensa compiler

* Add newlib __getreent stub for single-threaded bare-metal ESP32, eliminating linker warning

* Implement ESP32-S3 platform drivers using ESP-IDF LL APIs for all peripherals: system timing, GPIO, EXTI, UART, USB CDC, SPI, I2C, ADC, DShot/RMT, PWM motor/servo/beeper via LEDC, WS2812 LED strip, and flash config storage

* Fix ESP32 CI build: resolve Xtensa toolchain path during SDK registration

Move ESP_TOOLS_BIN resolution from ESP32S3.mk to tools.mk so that
PLATFORM_SDK_esp_idf_CC carries the full path when tools_check.mk
validates the compiler, matching the pattern used by the ARM SDK.

* Fix ESP32 review issues: add missing stdint.h include and remove duplicate source entries

- Add #include <stdint.h> to usb_msc_storage.h for uint32_t usage
- Remove drivers/dshot.c, drivers/bus_i2c_utils.c, and
  drivers/accgyro/accgyro_mpu.c from MCU_COMMON_SRC since they are
  already in COMMON_SRC (source.mk)

+++ src/main/flight/pid.c
@@ -1400,7 +1400,7 @@ void FAST_CODE pidController(const pidProfile_t *pidProfile, timeUs_t currentTim
                 dMaxMultiplier = pt2FilterApply(&pidRuntime.dMaxLowpass[axis], dMaxMultiplier);
                 // limit the gain to the fraction that DMax is greater than Min
                 dMaxMultiplier = MIN(dMaxMultiplier, pidRuntime.dMaxPercent[axis]);
-                if (debugMode == DEBUG_D_MAX && axis == gyro.gyroDebugAxis) {
+                if (debugMode == DEBUG_D_MAX && (int)axis == gyro.gyroDebugAxis) {
                     DEBUG_SET(DEBUG_D_MAX, 0, lrintf(dMaxGyroFactor * 100));
                     DEBUG_SET(DEBUG_D_MAX, 1, lrintf(dMaxSetpointFactor * 100));
                     DEBUG_SET(DEBUG_D_MAX, 2, lrintf(pidRuntime.pidCoefficient[axis].Kd * dMaxMultiplier * 10 / DTERM_SCALE)); // effective Kd after Dmax boost
